Baseline custom calendar component finished
