+++ client/src/AppointmentsPage/Calendar/components/CustomCalendar/Appointment.tsx
@@ -3,16 +3,30 @@ import { Stack, Typography, Box } from "@mui/material";
 
 export default function CustomAppointment({
   appointment,
+  onEventClick,
 }: {
   appointment: Appointment;
+  onEventClick?: (e: {
+    originalEvent: React.MouseEvent;
+    e: Appointment;
+  }) => void;
 }) {
   return (
-    <Box>
+    <Box
+      onMouseUp={(event) =>
+        onEventClick &&
+        onEventClick({
+          originalEvent: event,
+          e: appointment,
+        })
+      }
+    >
       <Stack>
         <Typography sx={{ fontSize: "0.8rem" }}>{appointment.name}</Typography>
         {appointment.services.map((service) => {
           return (
             <Typography
+              key={service}
               sx={{
                 fontSize: "0.7rem",
               }}
